third_party/WebKit/LayoutTests/fast/dom/shadow/wheel-event-on-input-in-shadow-dom.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/event-sender-util.js"></script>
<script src="../../../resources/js-test.js"></script>
</head>
<body>

<div id="host"></div>
<pre id="console"></pre>

<script>
function createInput(name) {
    var div = document.createElement('input');
    div.style.width = '100px';
    div.style.height = '100px';

    div.addEventListener('mousewheel', function(e) {
        debug('PASS: mousewheel event is fired.');
    });

    return div;
}

var shadowRoot = host.createShadowRoot();
var shadowDiv = createInput('shadow');
shadowRoot.appendChild(shadowDiv);

mouseMoveToElem(shadowDiv);
eventSender.mouseScrollBy(0, 1);

var successfullyParsed = true;
</script>

</body>
</html>
