LayoutTests/fast/dom/shadow/wheel-event-on-input-in-shadow-dom-expected.txt

PASS: mousewheel event is fired.
PASS successfullyParsed is true

TEST COMPLETE
